+++ src/Persistence/Windows/Migrations/SchemaMigrator.h
@@ -42,6 +42,7 @@ enum class SchemaLayout
     CentralVersion7,
     CentralVersion8,
     CentralVersion9,
+    CentralVersion10,
     ProjectVersion1,
     ProjectVersion2,
     ProjectVersion3,